# course-project/docs/feynman-course-apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Feynman Course API",
    "version": "0.1.0",
    "description": "课程版 Feynman 学习工作台 API。导入 Apifox 后，请先调用注册或登录接口获取 token，再在鉴权设置中配置 Bearer Token。"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "本地课程版服务端"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录或注册响应中的 token。请求头格式：Authorization: Bearer <token>"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string", "description": "错误信息", "example": "知识点不存在" }
        }
      },
...
      "FeynmanEvaluation": {
        "type": "object",
        "required": ["score", "evaluation", "polishedText", "strengths", "weaknesses", "provider"],
        "properties": {
          "score": { "type": "integer", "minimum": 0, "maximum": 100, "example": 82 },
          "evaluation": { "type": "string", "maxLength": 2000, "example": "解释覆盖了核心机制，下一次可以补充一个具体例子。" },
          "polishedText": { "type": "string", "maxLength": 30000, "example": "向量数据库把内容转换成向量，再通过相似度检索相关信息。" },
          "strengths": { "type": "array", "maxItems": 4, "items": { "type": "string" }, "example": ["包含核心定义"] },
          "weaknesses": { "type": "array", "maxItems": 4, "items": { "type": "string" }, "example": ["补充一个具体例子"] },
          "provider": { "type": "string", "example": "mock" }
        }
      },
      "FeynmanAttempt": {
        "allOf": [
          { "$ref": "#/components/schemas/FeynmanEvaluation" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "string", "readOnly": true },
              "knowledgeId": { "type": "string", "readOnly": true },
              "transcript": { "type": "string", "maxLength": 30000 },
              "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
              "updatedAt": { "type": "string", "format": "date-time", "readOnly": true }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/feynman/transcribe": {
      "post": {
        "tags": ["Feynman"],
        "summary": "上传费曼复述音频并转录",
        "description": "课程版服务端使用百度智能云短语音识别。前端会把浏览器录音转换为单声道 16kHz WAV；单段建议不超过 60 秒，音频大小上限为 10MB。",
        "operationId": "transcribeFeynman",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "type": "object", "required": ["knowledgeId", "audio"], "properties": { "knowledgeId": { "type": "string", "example": "66d123456789012345678902" }, "audio": { "type": "string", "format": "binary", "description": "单声道 16kHz WAV，建议不超过 60 秒" } }, "example": { "knowledgeId": "66d123456789012345678902" } }
            }
          }
        },
        "responses": {
          "200": { "description": "百度语音识别成功", "content": { "application/json": { "schema": { "type": "object", "properties": { "transcript": { "type": "string" }, "provider": { "type": "string", "example": "baidu-asr" }, "durationMs": { "type": "integer", "nullable": true, "example": null } } } } } },
          "400": { "description": "没有音频文件", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "百度语音识别凭证未配置", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "知识点不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/feynman/evaluate": {
      "post": {
        "tags": ["Feynman"],
        "summary": "评价费曼复述",
        "operationId": "evaluateFeynman",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["knowledgeId", "transcript"], "properties": { "knowledgeId": { "type": "string", "example": "66d123456789012345678902" }, "transcript": { "type": "string", "minLength": 1, "maxLength": 30000, "example": "向量数据库把内容转成向量，然后通过相似度找到相关信息。" } } } } } },
        "responses": {
          "200": { "description": "评价成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FeynmanEvaluation" } } } },
          "400": { "description": "转录内容为空", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "知识点不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "502": { "description": "AI Provider 请求失败或返回格式不正确", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/feynman/attempts": {
      "post": {
        "tags": ["Feynman"],
        "summary": "保存一次费曼复述（支持幂等）",
        "description": "clientAttemptId 用于防止网络重试或重复点击产生重复复述记录；相同用户和相同 clientAttemptId 再次提交时返回已有记录。",
        "operationId": "createFeynmanAttempt",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["knowledgeId", "transcript", "polishedText", "score", "evaluation"], "properties": { "knowledgeId": { "type": "string", "example": "66d123456789012345678902" }, "clientAttemptId": { "type": "string", "maxLength": 100, "description": "客户端生成的本次复述唯一标识，建议使用 UUID" }, "transcript": { "type": "string", "maxLength": 30000 }, "polishedText": { "type": "string", "maxLength": 30000 }, "score": { "type": "integer", "minimum": 0, "maximum": 100 }, "evaluation": { "type": "string", "maxLength": 2000 }, "strengths": { "type": "array", "items": { "type": "string" } }, "weaknesses": { "type": "array", "items": { "type": "string" } } } } } } },
        "responses": {
          "200": { "description": "重复请求，返回已存在的复述记录", "content": { "application/json": { "schema": { "type": "object", "properties": { "attempt": { "$ref": "#/components/schemas/FeynmanAttempt" }, "duplicate": { "type": "boolean", "example": true } } } } } },
          "201": { "description": "保存成功", "content": { "application/json": { "schema": { "type": "object", "properties": { "attempt": { "$ref": "#/components/schemas/FeynmanAttempt" } } } } } },
          "404": { "description": "知识点不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/feynman/attempts/{knowledgeId}": {
      "get": {
        "tags": ["Feynman"],
        "summary": "查询知识点的费曼复述记录",
        "operationId": "listFeynmanAttempts",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "knowledgeId", "in": "path", "required": true, "schema": { "type": "string" }, "example": "66d123456789012345678902" }],
        "responses": {
          "200": { "description": "复述记录列表", "content": { "application/json": { "schema": { "type": "object", "properties": { "attempts": { "type": "array", "items": { "$ref": "#/components/schemas/FeynmanAttempt" } } } } } } },
          "404": { "description": "知识点不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}